🧊 Added shared Redis config
